Add interface for NamespaceRateLimitInterceptor
