+++ client/history/client_test.go
@@ -36,7 +36,7 @@ import (
 	"go.temporal.io/server/common/dynamicconfig"
 	"go.temporal.io/server/common/log"
 	"go.temporal.io/server/common/membership"
-	"go.temporal.io/server/internal/nettest"
+	"go.temporal.io/server/common/nettest"
 	"go.uber.org/mock/gomock"
 	"google.golang.org/grpc"
 )
