+++ common/membership/grpc_resolver_test.go
@@ -32,8 +32,8 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"go.temporal.io/server/common/nettest"
 	"go.temporal.io/server/common/primitives"
-	"go.temporal.io/server/internal/nettest"
 	"go.uber.org/mock/gomock"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
