+++ service/frontend/fx_test.go
@@ -41,9 +41,9 @@ import (
 	"go.temporal.io/server/common/metrics"
 	"go.temporal.io/server/common/metrics/metricstest"
 	"go.temporal.io/server/common/namespace"
+	"go.temporal.io/server/common/nettest"
 	"go.temporal.io/server/common/primitives"
 	"go.temporal.io/server/common/rpc"
-	"go.temporal.io/server/internal/nettest"
 	"go.uber.org/mock/gomock"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
